+++ BrotBoxEngine/BBE/RenderManager.h
@@ -58,7 +58,7 @@ namespace bbe
 		virtual void fillRect2D(const Rectangle& rect, float rotation, FragmentShader* shader) = 0;
 		virtual void fillCircle2D(const Circle& circle) = 0;
 		virtual void drawImage2D(const Rectangle& rect, const Image& image, float rotation) = 0;
-		virtual void fillVertexIndexList2D(const uint32_t* indices, uint32_t amountOfIndices, const bbe::Vector2* vertices, size_t amountOfVertices, const bbe::Vector2& pos, const bbe::Vector2 &scale) = 0;
+		virtual void fillVertexIndexList2D(const uint32_t* indices, size_t amountOfIndices, const bbe::Vector2* vertices, size_t amountOfVertices, const bbe::Vector2& pos, const bbe::Vector2 &scale) = 0;
 
 		void setFillMode3D(bbe::FillMode fm);
 		bbe::FillMode getFillMode3D();
