+++ BrotBoxEngine/Vector2.cpp
@@ -9,5 +9,5 @@ uint32_t bbe::hash(const Vector2i& t)
 template<>
 uint32_t bbe::hash(const Vector2& t)
 {
-	return (((int32_t)t.x) ^ ((int32_t)t.y)) + t.x * 7001.f + t.y * 17.f;
+	return uint32_t((((int32_t)t.x) ^ ((int32_t)t.y)) + t.x * 7001.f + t.y * 17.f);
 }
